+++ packages/gamut-labs/src/AppHeader/shared.tsx
@@ -1,8 +1,7 @@
-import { IconButton } from '@codecademy/gamut';
-import { css } from '@codecademy/gamut-styles';
-import styled from '@emotion/styled';
 import { AnimatePresence, motion, Variants } from 'framer-motion';
 import React from 'react';
+
+export const appHeaderMobileBreakpoint = 'lg' as const;
 
 export type AnimatedHeaderZoneProps = {
   visible?: boolean;
@@ -30,11 +29,3 @@ export const AnimatedHeaderZone: React.FC<AnimatedHeaderZoneProps> = ({
     </AnimatePresence>
   ) : null;
 };
-
-export const HeaderIconButton = styled(IconButton)(
-  css({
-    '&:hover': {
-      background: `none`,
-    },
-  })
-);