+++ packages/gamut/src/Button/variants.ts
@@ -29,8 +29,16 @@ export const textButtonVariants = templateVariants(
   buttonVariants,
   (variant) => ({
     borderColor: 'transparent',
-    color: variant,
-    [ButtonSelectors.OUTLINE]: { borderColor: variant },
+    color: variant === 'interface' ? 'text' : variant,
+    [ButtonSelectors.HOVER]: {
+      color: variant,
+    },
+    [ButtonSelectors.FOCUS_VISIBLE]: {
+      color: variant,
+    },
+    [ButtonSelectors.OUTLINE]: {
+      borderColor: variant,
+    },
     [ButtonSelectors.ACTIVE]: {
       color: 'text',
     },
@@ -93,6 +101,7 @@ export const ctaButtonVariants = templateVariants(['primary'], (variant) => ({
 export const sizeVariants = variant({
   prop: 'size',
   defaultVariant: 'normal',
+  base: { fontWeight: 'title' },
   variants: {
     normal: {
       fontSize: 16,