+++ packages/gamut-patterns/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.8.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.8.2...@codecademy/gamut-patterns@0.8.3) (2022-02-18)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
+
+
+
+
+### [0.8.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.8.1...@codecademy/gamut-patterns@0.8.2) (2022-02-01)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
+
+
+
+
+### [0.8.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.8.0...@codecademy/gamut-patterns@0.8.1) (2022-01-31)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
+
+
+
+
 ## [0.8.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.7.3...@codecademy/gamut-patterns@0.8.0) (2022-01-25)