+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,158 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [54.3.4](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.3.3...@codecademy/styleguide@54.3.4) (2022-02-18)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.3.3](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.3.2...@codecademy/styleguide@54.3.3) (2022-02-16)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.3.2](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.3.1...@codecademy/styleguide@54.3.2) (2022-02-09)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.3.1](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.3.0...@codecademy/styleguide@54.3.1) (2022-02-08)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [54.3.0](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.2.1...@codecademy/styleguide@54.3.0) (2022-02-08)
+
+
+### Features
+
+* **Loading:** add dark mode ([7f5c97d](https://github.com/Codecademy/gamut/commit/7f5c97d746f9b293700ff230beba199c7178a19f))
+
+
+### Bug Fixes
+
+* Markdown details element styles ([5aa71ad](https://github.com/Codecademy/gamut/commit/5aa71ad077bfaf17c3ee154d1a25bb130fdc66d5))
+
+
+
+### [54.2.1](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.2.0...@codecademy/styleguide@54.2.1) (2022-02-04)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [54.2.0](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.1.2...@codecademy/styleguide@54.2.0) (2022-02-04)
+
+
+### Features
+
+* **Pagination:** controlled pagination ([9601492](https://github.com/Codecademy/gamut/commit/960149259c14ce6dd02368025cb4ad0d8f7db990))
+
+
+
+### [54.1.2](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.1.1...@codecademy/styleguide@54.1.2) (2022-02-04)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.1.1](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.1.0...@codecademy/styleguide@54.1.1) (2022-02-03)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [54.1.0](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.0.8...@codecademy/styleguide@54.1.0) (2022-02-02)
+
+
+### Features
+
+* **Pagination:** add pagination component ([5016301](https://github.com/Codecademy/gamut/commit/50163013e66f18769542a3aeef07781afae08f8b))
+
+
+
+### [54.0.8](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.0.7...@codecademy/styleguide@54.0.8) (2022-02-02)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.0.7](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.0.6...@codecademy/styleguide@54.0.7) (2022-02-01)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.0.6](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.0.5...@codecademy/styleguide@54.0.6) (2022-01-31)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.0.5](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.0.4...@codecademy/styleguide@54.0.5) (2022-01-28)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.0.4](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.0.3...@codecademy/styleguide@54.0.4) (2022-01-28)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.0.3](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.0.2...@codecademy/styleguide@54.0.3) (2022-01-27)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.0.2](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.0.1...@codecademy/styleguide@54.0.2) (2022-01-27)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [54.0.1](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@54.0.0...@codecademy/styleguide@54.0.1) (2022-01-26)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
 ## [54.0.0](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@53.1.1...@codecademy/styleguide@54.0.0) (2022-01-25)